requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2==2.9.10
Pygments==2.19.1
python-dotenv==1.0.1
rich==13.9.4